change to ADIOS2_USE_KVCACHE from ADIOS2_USE_Cache to be consistent in naming (#4250)

+++ source/adios2/toolkit/kvcache/KVCacheCommon.cpp
@@ -119,7 +119,7 @@ void KVCacheCommon::KeyPrefixExistence(const std::string &key_prefix,
     }
     else
     {
-        for (int i = 0; i < m_redisReply->elements; i++)
+        for (size_t i = 0; i < m_redisReply->elements; i++)
         {
             keys.insert(m_redisReply->element[i]->str);
         }
